changed tape flag for solved level to tape property bit
This change has the advantage that some more property bits can be
stored in tape files.

It has the disadvantage that previous program versions may treat a
tape as a solution tape when it really isn't (but instead has other
property bits set). As this is only used for the "dump tape" command
on the command line, this potential problem can safely be ignored.

+++ src/game.c
@@ -17238,7 +17238,7 @@ static void GameUndo(int steps)
   if (!CheckEngineSnapshotList())
     return;
 
-  int tape_property_bits = tape.property_bits;
+  unsigned short tape_property_bits = tape.property_bits;
 
   LoadEngineSnapshot_Undo(steps);
 
@@ -17252,7 +17252,7 @@ static void GameRedo(int steps)
   if (!CheckEngineSnapshotList())
     return;
 
-  int tape_property_bits = tape.property_bits;
+  unsigned short tape_property_bits = tape.property_bits;
 
   LoadEngineSnapshot_Redo(steps);
